requirements.txt
streamlit
yt_dlp
youtube_transcript_api
google.generativeai
